server/requirements/test.txt
#
# This file is autogenerated by pip-compile
# Make changes in test.in, then run this to update:
#
#    pip-compile test.in
#
click==6.2                # via pip-tools
coverage==4.0.3
defusedxml==0.4.1         # via python3-openid
django-allauth==0.24.1
django-autoslug==1.9.3
django-braces==1.8.1
django-crispy-forms==1.5.2
django-environ==0.4.0
django-filter==0.11.0
django-floppyforms==1.5.2
django-model-utils==2.3.1
django-redis==4.3.0
django-secure==1.0.1
django-test-plus==1.0.11
django==1.8.7
djangorestframework-jwt==1.7.2
djangorestframework==3.3.1
factory-boy==2.6.0
fake-factory==0.5.3
first==2.0.1              # via pip-tools
flake8==2.5.0
mccabe==0.3.1             # via flake8
oauthlib==1.0.3           # via requests-oauthlib
pep8==1.5.7               # via flake8
pillow==3.0.0
pip-tools==1.1.6
psycopg2==2.6.1
pyflakes==1.0.0           # via flake8
pyjwt==1.4.0              # via djangorestframework-jwt
python3-openid==3.0.9     # via django-allauth
pytz==2015.7
pyyaml==3.11
redis==2.10.5
requests-oauthlib==0.5.0  # via django-allauth
requests==2.8.1           # via django-allauth, requests-oauthlib
six==1.10.0               # via django-braces, django-environ, pip-tools, unicode-slugify
unicode-slugify==0.1.3
unidecode==0.4.18         # via unicode-slugify
wheel==0.26.0
whitenoise==2.0.6
